Separate out handleRecording into its own function.
Simple refactoring. Created a new 'useRecording' hook to separate out complex logic that was in createPrayer file.
